refactor: add interfaces and enum
