+++ src/controllers/checkUserValidate.ts
@@ -1,4 +1,4 @@
-import { IUsers } from "../server.js";
+import {HEADER_CONTENT_TYPE, IUsers, statusCode} from '../interfaces/interfaces.js';
 
 export function checkUser(user:IUsers):boolean {
     const id = user.hasOwnProperty("id");
