+++ src/controllers/getAllUsers.ts
@@ -1,9 +1,9 @@
 import { IncomingMessage, ServerResponse } from "http";
-import { IUsers } from "../server.js";
+import {HEADER_CONTENT_TYPE, IUsers, statusCode} from '../interfaces/interfaces.js';
 
 
 export function getAllUsers(req:IncomingMessage, res:ServerResponse, users:IUsers[]):IUsers[] {
-    res.writeHead(200, { 'Content-Type': 'application/json' });
+    res.writeHead(statusCode.success, HEADER_CONTENT_TYPE);
     res.end(JSON.stringify(users));
     return users;
 }
